\name{exponentialDistribution}
\alias{exponentialDistribution}

\title{
Exponential Distribution
}
\description{

  Paramters: λ > 0

  Value domain: [0, Inf)

  Density function: f(x) = λe^(−λx); x ≥ 0

  Distribution: F(x) = 1 - e^(-λx)

  An exponential distribution modelates waiting time,
  with the aditional property that it lacks memory.
}
\usage{
  probability of an unstable isotope to decompose.
  In this case λ is linked to the half-life of the isotope.

  probability of a taxi to arrive in x minutes.
  Here λ is 1/mean of a taxi to pass in a minute.
}
\details{
  the property of lacking memory means that if the probability of an event to happen in n units of time is "p", then the probability of the event happening in the next n units of time is the same "p""
}
\source{
  Materiale Curs
}
