\name{normalDistribution}
\alias{normalDistribution}
\alias{gaussianDistribution}

\title{
  Normal Distribution
}
\description{
  Parameters: µ ∈ R, σ > 0

  Value domain: R

  Density function: f(x) = (1/σ√(2*π)) * e^(-(x - µ)^2/2σ^2)

  Repartition: F(x) have no formula, we use tables or command like pnorm in R to get the value
}
\usage{
  Calculate error value, inteligence, height, means of data batches
}
\details{
  The normal distribution N(0, 1) have a mean of 0 and diseprsion of 1.
}
\source{
  Materiale curs
}

