\name{paretoDistribution}
\alias{paretoDistribution}

\title{
  Pareto Distribution
}
\description{
  Parameters: m > 0, α > 0

  Value Domain: [m, Inf)

  Density function: f(x) = αm^α / x^(α + 1); x >= m

  Distribution: F(x) = 1 - (m^α / x^α); x>= m
}
\usage{
  Pareto distribution models a power rule, where the probability of an event varies as a power of an attribute of an event.

  Some examples:
    -size of meteorites, population size, Hard disk drive error rates, user play time on differet games on Steam, etc.
}
\source{
  Materiale curs, \href{https://en.wikipedia.org/wiki/Pareto_distribution}{wikipedia}
}
