\name{uniformDistribution}
\alias{uniformDistribution}

\title{
  Uniform distribution
}
\description{
  All results in the domain have equal probability.

  Value domain: [a, b]

  Density function: f(x) = 1/(b - a); a <= x <= b

  Distribution: F(x) = (x - a)/ (b - a)

  A continuous uniform distribution usually comes in a rectangular shape.

  A good example of a continuous uniform distribution is an idealized random number generator.

  With continuous uniform distribution,
  just like discrete uniform distribution,
  every variable has an equal chance of happening.
  However, there is an infinite number of points that can exist.
}
\source{
  Materiale Curs, \href{https://corporatefinanceinstitute.com/resources/knowledge/other/uniform-distribution/}{CFI}
}

